+++ rest-ehr-scape/src/main/java/org/ehrbase/rest/ehrscape/responsedata/TemplateResponseData.java
@@ -19,7 +19,8 @@
 package org.ehrbase.rest.ehrscape.responsedata;
 
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement;
-import org.ehrbase.response.ehrscape.WebTemplate;
+import org.ehrbase.webtemplate.WebTemplate;
+
 
 @JacksonXmlRootElement
 public class TemplateResponseData extends ActionRestResponseData {