+++ service/src/main/java/org/ehrbase/service/KnowledgeCacheService.java
@@ -22,45 +22,22 @@
  */
 package org.ehrbase.service;
 
-import static org.ehrbase.configuration.CacheConfiguration.OPERATIONAL_TEMPLATE_CACHE;
-import static org.ehrbase.configuration.CacheConfiguration.QUERY_CACHE;
-
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
-import java.util.UUID;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.stream.Collectors;
-import java.util.stream.StreamSupport;
-
-import javax.annotation.PostConstruct;
-import javax.annotation.PreDestroy;
-import javax.cache.Cache;
-import javax.cache.CacheManager;
-
-import org.apache.commons.collections4.MapUtils;
+import com.google.common.util.concurrent.MoreExecutors;
 import org.apache.xmlbeans.XmlException;
 import org.ehrbase.api.exception.InternalServerException;
 import org.ehrbase.api.exception.InvalidApiParameterException;
 import org.ehrbase.api.exception.StateConflictException;
 import org.ehrbase.aql.containment.JsonPathQueryResult;
-import org.ehrbase.aql.containment.OptJsonPath;
+import org.ehrbase.aql.containment.TemplateIdAqlTuple;
 import org.ehrbase.aql.containment.TemplateIdQueryTuple;
+import org.ehrbase.aql.sql.queryImpl.ItemInfo;
 import org.ehrbase.configuration.CacheConfiguration;
 import org.ehrbase.ehr.knowledge.I_KnowledgeCache;
 import org.ehrbase.ehr.knowledge.TemplateMetaData;
-import org.ehrbase.opt.OptVisitor;
-import org.ehrbase.opt.query.I_QueryOptMetaData;
-import org.ehrbase.opt.query.MapJson;
-import org.ehrbase.opt.query.QueryOptMetaData;
+import org.ehrbase.webtemplate.NodeId;
+import org.ehrbase.webtemplate.OPTParser;
+import org.ehrbase.webtemplate.WebTemplate;
+import org.ehrbase.webtemplate.WebTemplateNode;
 import org.openehr.schemas.v1.OBJECTID;
 import org.openehr.schemas.v1.OPERATIONALTEMPLATE;
 import org.openehr.schemas.v1.TEMPLATEID;
@@ -71,6 +48,36 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
+
+import javax.annotation.PostConstruct;
+import javax.annotation.PreDestroy;
+import javax.cache.Cache;
+import javax.cache.CacheManager;
+import java.io.ByteArrayInputStream;
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
+import java.util.TreeSet;
+import java.util.UUID;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ThreadPoolExecutor;
+import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
+import java.util.stream.StreamSupport;
+
+import static org.ehrbase.configuration.CacheConfiguration.FIELDS_CACHE;
+import static org.ehrbase.configuration.CacheConfiguration.MULTI_VALUE_CACHE;
+import static org.ehrbase.configuration.CacheConfiguration.OPERATIONAL_TEMPLATE_CACHE;
+import static org.ehrbase.configuration.CacheConfiguration.QUERY_CACHE;
 
 /**
  * Look up and caching for archetypes, openEHR showTemplates and Operational Templates. Search in path defined as
@@ -101,39 +108,55 @@ public class KnowledgeCacheService implements I_KnowledgeCache, IntrospectServic
     private final TemplateStorage templateStorage;
     private final Cache<TemplateIdQueryTuple, JsonPathQueryResult> jsonPathQueryResultCache;
 
-    private Cache<String, OPERATIONALTEMPLATE> atOptCache;
-    private final Cache<UUID, I_QueryOptMetaData> queryOptMetaDataCache;
+    private final Cache<String, OPERATIONALTEMPLATE> atOptCache;
+    private final Cache<UUID, WebTemplate> webTemplateCache;
+    private final Cache<TemplateIdAqlTuple, ItemInfo> fieldCache;
 
-    //index
-    //template index with UUID (not used so far...)
-    private Map<UUID, String> idxCache = new ConcurrentHashMap<>();
+    private final Cache<String, List> multivaluedCache;
+
+    //index uuid to templateId
+    private Map<UUID, String> idxCacheUuidToTemplateId = new ConcurrentHashMap<>();
+    //index templateId to uuid
+    private Map<String, UUID> idxCacheTemplateIdToUuid = new ConcurrentHashMap<>();
 
     private Set<String> allTemplateId = new HashSet<>();
 
-    private Map<String, Set<String>> nodeIdsByTemplateIdMap = new HashMap<>();
-
 
     private final CacheManager cacheManager;
+    private final CacheConfiguration cacheConfiguration;
 
     @Value("${system.allow-template-overwrite:false}")
     private boolean allowTemplateOverwrite;
 
     @Autowired
-    public KnowledgeCacheService(
-            @Qualifier("templateDBStorageService") TemplateStorage templateStorage,
-            CacheManager cacheManager
-            ) {
+    public KnowledgeCacheService(@Qualifier("templateDBStorageService") TemplateStorage templateStorage, CacheManager cacheManager, CacheConfiguration cacheConfiguration) {
         this.templateStorage = templateStorage;
         this.cacheManager = cacheManager;
 
         atOptCache = cacheManager.getCache(OPERATIONAL_TEMPLATE_CACHE, String.class, OPERATIONALTEMPLATE.class);
-        queryOptMetaDataCache = cacheManager.getCache(CacheConfiguration.INTROSPECT_CACHE, UUID.class, I_QueryOptMetaData.class);
+        webTemplateCache = cacheManager.getCache(CacheConfiguration.INTROSPECT_CACHE, UUID.class, WebTemplate.class);
         jsonPathQueryResultCache = cacheManager.getCache(QUERY_CACHE, TemplateIdQueryTuple.class, JsonPathQueryResult.class);
+        fieldCache = cacheManager.getCache(FIELDS_CACHE, TemplateIdAqlTuple.class, ItemInfo.class);
+        multivaluedCache = cacheManager.getCache(MULTI_VALUE_CACHE, String.class, List.class);
+        this.cacheConfiguration = cacheConfiguration;
     }
 
     @PostConstruct
     public void init() {
         allTemplateId = listAllOperationalTemplates().stream().map(TemplateMetaData::getOperationaltemplate).map(OPERATIONALTEMPLATE::getTemplateId).map(OBJECTID::getValue).collect(Collectors.toSet());
+        listAllOperationalTemplates().stream().map(TemplateMetaData::getOperationaltemplate).forEach(this::putIntoCache);
+
+
+        if (cacheConfiguration.isPreBuildQueries()) {
+            ThreadPoolExecutor executor = (ThreadPoolExecutor) Executors.newFixedThreadPool(1);
+            ExecutorService executorService =
+                    MoreExecutors.getExitingExecutorService(executor,
+                            100, TimeUnit.MILLISECONDS);
+
+            executor.submit(() ->
+                    allTemplateId.forEach(this::precalculateQuerys));
+        }
+
     }
 
     @PreDestroy
@@ -148,71 +171,94 @@ public class KnowledgeCacheService implements I_KnowledgeCache, IntrospectServic
 
     @Override
     public String addOperationalTemplate(byte[] content) {
-        return addOperationalTemplate(content, false);
+        return addOperationalTemplateIntern(content, false);
     }
 
-    /**
-     * Creates a new or replaces an existing operational template. If the template does not exist it will be created.
-     * If there is already a template with the given id inside the content and either the configuration setting
-     * system.allow-template-overwrite or param overwrite is set to true the template will be replaced with the new
-     * content. Of none of these flags is set a conflict exception will be thrown.
-     *
-     * @param content - New template content to write / set
-     * @param overwrite - Allow overwrite of existing templates
-     * @return - New created template id
-     */
-    private String addOperationalTemplate(byte[] content, boolean overwrite) {
+    public String addOperationalTemplateIntern(byte[] content, boolean overwrite) {
 
-        OPERATIONALTEMPLATE template = parseTemplate(content);
+        InputStream inputStream = new ByteArrayInputStream(content);
+
+        TemplateDocument document;
+        try {
+            document = TemplateDocument.Factory.parse(inputStream);
+        } catch (XmlException | IOException e) {
+            throw new InvalidApiParameterException(e.getMessage());
+        }
+        OPERATIONALTEMPLATE template = document.getTemplate();
+
+        if (template == null) {
+            throw new InvalidApiParameterException("Could not parse input template");
+        }
+
+        if (template.getConcept() == null || template.getConcept().isEmpty())
+            throw new IllegalArgumentException("Supplied template has nil or empty concept");
+
+        if (template.getDefinition() == null || template.getDefinition().isNil())
+            throw new IllegalArgumentException("Supplied template has nil or empty definition");
+
+        if (template.getDescription() == null || !template.getDescription().validate())
+            throw new IllegalArgumentException("Supplied template has nil or empty description");
 
         //get the filename from the template template Id
         Optional<TEMPLATEID> filenameOptional = Optional.ofNullable(template.getTemplateId());
         String templateId = filenameOptional.orElseThrow(() -> new InvalidApiParameterException("Invalid template input content")).getValue();
 
-        // pre-check: if already existing and overwrite is forbidden throw proper exception
+
+        // pre-check: if already existing throw proper exception
         if (!allowTemplateOverwrite && !overwrite && retrieveOperationalTemplate(templateId).isPresent()) {
             throw new StateConflictException("Operational template with this template ID already exists");
         }
 
         templateStorage.storeTemplate(template);
 
+
         invalidateCache(template);
 
-        atOptCache.put(templateId, template);
-        idxCache.put(UUID.fromString(template.getUid().getValue()), templateId);
-        allTemplateId.add(templateId);
+        putIntoCache(template);
+        if (cacheConfiguration.isPreBuildQueries()) {
+            ThreadPoolExecutor executor = (ThreadPoolExecutor) Executors.newFixedThreadPool(1);
+            ExecutorService executorService =
+                    MoreExecutors.getExitingExecutorService(executor,
+                            100, TimeUnit.MILLISECONDS);
+
+            executor.submit(() -> precalculateQuerys(templateId));
+        }
 
         //retrieve the template Id for this new entry
         return template.getTemplateId().getValue();
     }
 
-    public String adminUpdateOperationalTemplate(byte[] content) {
+    private void putIntoCache(OPERATIONALTEMPLATE template) {
+        String templateId = template.getTemplateId().getValue();
+        atOptCache.put(templateId, template);
+        idxCacheUuidToTemplateId.put(UUID.fromString(template.getUid().getValue()), templateId);
+        idxCacheTemplateIdToUuid.put(templateId, UUID.fromString(template.getUid().getValue()));
+        allTemplateId.add(templateId);
 
-        OPERATIONALTEMPLATE template = parseTemplate(content);
-
-        String templateId = Optional.ofNullable(
-                template.getTemplateId())
-                .orElseThrow(() -> new InvalidApiParameterException("Invalid template input content"))
-                .getValue();
-
-        // Replace template
-        templateStorage.adminUpdateTemplate(template);
-
-        // Refresh template caches
-        invalidateCache(template);
-        atOptCache.replace(templateId, template);
-        idxCache.replace(UUID.fromString(template.getUid().getValue()), templateId);
-
-        return template.xmlText();
+        getQueryOptMetaData(templateId);
     }
 
+    private void precalculateQuerys(String templateId) {
+
+        getQueryOptMetaData(templateId).findAllContainmentCombinations()
+                .stream()
+                .filter(s -> !s.isEmpty())
+                .filter(s -> s.size() <= cacheConfiguration.getPreBuildQueriesDepth())
+                .forEach(s -> resolveForTemplate(templateId, s));
+
+
+    }
+
+    public String adminUpdateOperationalTemplate(byte[] content) {
+
+        return addOperationalTemplateIntern(content, true);
+    }
 
     // invalidates some derived caches like the queryOptMetaDataCache which depend on the template
     private void invalidateCache(OPERATIONALTEMPLATE template) {
 
         //invalidate the cache for this template
-        queryOptMetaDataCache.remove(UUID.fromString(template.getUid().getValue()));
-        atOptCache.remove(template.getUid().getValue());
+        webTemplateCache.remove(UUID.fromString(template.getUid().getValue()));
         Set<TemplateIdQueryTuple> collect = StreamSupport.stream(jsonPathQueryResultCache.spliterator(), true)
                 .map(Cache.Entry::getKey)
 
@@ -220,6 +266,7 @@ public class KnowledgeCacheService implements I_KnowledgeCache, IntrospectServic
                 .collect(Collectors.toSet());
         jsonPathQueryResultCache.removeAll(collect);
     }
+
 
     @Override
     public List<TemplateMetaData> listAllOperationalTemplates() {
@@ -247,29 +294,59 @@ public class KnowledgeCacheService implements I_KnowledgeCache, IntrospectServic
         return retrieveOperationalTemplate(key);
     }
 
-    private String findTemplateIdByUuid(UUID uuid) {
-        String key = idxCache.get(uuid);
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public boolean deleteOperationalTemplate(OPERATIONALTEMPLATE template) {
+        // Remove template from storage
+        boolean deleted = this.templateStorage.deleteTemplate(template.getTemplateId().getValue());
 
-        if (key == null) {
-            key = listAllOperationalTemplates()
+        if (deleted) {
+            // Remove template from caches
+            invalidateCache(template);
+        }
+
+        return deleted;
+    }
+
+    private String findTemplateIdByUuid(UUID uuid) {
+        String templateId = idxCacheUuidToTemplateId.get(uuid);
+
+        if (templateId == null) {
+            templateId = listAllOperationalTemplates()
                     .stream()
                     .filter(t -> t.getErrorList().isEmpty())
                     .filter(t -> t.getOperationaltemplate().getUid().getValue().equals(uuid.toString()))
                     .map(t -> t.getOperationaltemplate().getTemplateId().getValue())
                     .findFirst()
                     .orElse(null);
+            idxCacheUuidToTemplateId.put(uuid, templateId);
         }
-        return key;
+
+        return templateId;
+    }
+
+    private UUID findUuidByTemplateId(String templateId) {
+        UUID uuid = idxCacheTemplateIdToUuid.get(templateId);
+        if (uuid == null) {
+            uuid = UUID.fromString(retrieveOperationalTemplate(templateId)
+                    .orElseThrow()
+                    .getUid()
+                    .getValue());
+            idxCacheTemplateIdToUuid.put(templateId, uuid);
+        }
+        return uuid;
     }
 
 
     @Override
-    public I_QueryOptMetaData getQueryOptMetaData(UUID uuid) {
+    public WebTemplate getQueryOptMetaData(UUID uuid) {
 
-        final I_QueryOptMetaData retval;
+        final WebTemplate retval;
 
-        if (queryOptMetaDataCache.containsKey(uuid))
-            retval = queryOptMetaDataCache.get(uuid);
+        if (webTemplateCache.containsKey(uuid))
+            retval = webTemplateCache.get(uuid);
         else {
             retval = buildAndCacheQueryOptMetaData(uuid);
         }
@@ -277,33 +354,13 @@ public class KnowledgeCacheService implements I_KnowledgeCache, IntrospectServic
     }
 
     @Override
-    public I_QueryOptMetaData getQueryOptMetaData(String templateId) {
+    public WebTemplate getQueryOptMetaData(String templateId) {
 
-        //get the matching template if any
-
-
-        Optional<OPERATIONALTEMPLATE> operationaltemplate = retrieveOperationalTemplate(templateId);
-
-        if (operationaltemplate.isPresent())
-            return getQueryOptMetaData(UUID.fromString(operationaltemplate.get().getUid().getValue()));
-        else {
-
-            Optional<OPERATIONALTEMPLATE> cachedOpt = Optional.empty();
-            try {
-                cachedOpt = retrieveOperationalTemplate(templateId);
-            } catch (Exception e) {
-                log.warn(e.getMessage(), e);
-            }
-            if (cachedOpt.isPresent()) {
-                UUID uuid = UUID.fromString(cachedOpt.get().getUid().getValue());
-                return getQueryOptMetaData(uuid);
-            } else
-                throw new IllegalArgumentException("Could not retrieve  knowledgeCacheService.getKnowledgeCache() cache for template id:" + templateId);
-        }
+        return getQueryOptMetaData(findUuidByTemplateId(templateId));
     }
 
-    private I_QueryOptMetaData buildAndCacheQueryOptMetaData(UUID uuid) {
-        I_QueryOptMetaData retval;
+    private WebTemplate buildAndCacheQueryOptMetaData(UUID uuid) {
+        WebTemplate retval;
         Optional<OPERATIONALTEMPLATE> operationaltemplate = Optional.empty();
         try {
             operationaltemplate = retrieveOperationalTemplate(uuid);
@@ -318,17 +375,16 @@ public class KnowledgeCacheService implements I_KnowledgeCache, IntrospectServic
         return retval;
     }
 
-    private I_QueryOptMetaData buildAndCacheQueryOptMetaData(OPERATIONALTEMPLATE operationaltemplate) {
-        log.info("Updating getQueryOptMetaData cache for template: {}", operationaltemplate.getTemplateId().getValue());
-        final I_QueryOptMetaData visitor;
+    private WebTemplate buildAndCacheQueryOptMetaData(OPERATIONALTEMPLATE operationaltemplate) {
+        log.info("Updating WebTemplate cache for template: {}", operationaltemplate.getTemplateId().getValue());
+        final WebTemplate visitor;
         try {
-            Map map = new OptVisitor().traverse(operationaltemplate);
-            visitor = QueryOptMetaData.getInstance(new MapJson(map).toJson());
+            visitor = new OPTParser(operationaltemplate).parse();
         } catch (Exception e) {
             throw new InternalServerException(e.getMessage(), e);
         }
 
-        queryOptMetaDataCache.put(UUID.fromString(operationaltemplate.getUid().getValue()), visitor);
+        webTemplateCache.put(UUID.fromString(operationaltemplate.getUid().getValue()), visitor);
         return visitor;
     }
 
@@ -343,26 +399,9 @@ public class KnowledgeCacheService implements I_KnowledgeCache, IntrospectServic
         OPERATIONALTEMPLATE operationaltemplate = templateStorage.readOperationaltemplate(filename).orElse(null);
         if (operationaltemplate != null) {
             atOptCache.put(filename, operationaltemplate);      // manual putting into cache (actual opt cache and then id cache)
-            idxCache.put(UUID.fromString(operationaltemplate.getUid().getValue()), filename);
+            idxCacheUuidToTemplateId.put(UUID.fromString(operationaltemplate.getUid().getValue()), filename);
         }
         return operationaltemplate;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    public boolean deleteOperationalTemplate(OPERATIONALTEMPLATE template) {
-        // Remove template from storage
-        boolean deleted = this.templateStorage.deleteTemplate(template.getTemplateId().getValue());
-
-        if (deleted) {
-            // Remove template from caches
-            this.atOptCache.remove(template.getTemplateId().getValue());
-            this.idxCache.remove(UUID.fromString(template.getUid().getValue()));
-        }
-
-        return deleted;
     }
 
     public int deleteAllOperationalTemplates() {
@@ -372,31 +411,60 @@ public class KnowledgeCacheService implements I_KnowledgeCache, IntrospectServic
         if (templateList.isEmpty()) {
             return 0;
         }
-        int deleted = this.templateStorage.adminDeleteAllTemplates(templateList);
+        int deleted = 0;
+        for (TemplateMetaData metaData : templateList) {
 
-        // Clear cache
-        this.atOptCache.clear();
-        this.idxCache.clear();
+            if (deleteOperationalTemplate(metaData.getOperationaltemplate())) {
+                deleted++;
+            }
+        }
 
         return deleted;
     }
-    
-    @Override
-    public boolean containsNodeIds(String templateId, Collection<String> nodeIds) {
-        Set<String> templateNodeIds = nodeIdsByTemplateIdMap.computeIfAbsent(templateId, t -> getQueryOptMetaData(t).getAllNodeIds());
-        return templateNodeIds.containsAll(nodeIds);
-    }
+
 
     @Override
-    public JsonPathQueryResult resolveForTemplate(String templateId, String jsonQueryExpression) {
-        TemplateIdQueryTuple key = new TemplateIdQueryTuple(templateId, jsonQueryExpression);
+    public JsonPathQueryResult resolveForTemplate(String templateId, Collection<NodeId> nodeIds) {
+        TemplateIdQueryTuple key = new TemplateIdQueryTuple(templateId, nodeIds);
 
         JsonPathQueryResult jsonPathQueryResult = jsonPathQueryResultCache.get(key);
-
         if (jsonPathQueryResult == null) {
-            Map<String, Object> evaluate = new OptJsonPath(this).evaluate(templateId, jsonQueryExpression);
-            if (!MapUtils.isEmpty(evaluate)) {
-                jsonPathQueryResult = new JsonPathQueryResult(templateId, evaluate);
+
+
+            WebTemplate webTemplate = getQueryOptMetaData(templateId);
+            List<WebTemplateNode> webTemplateNodeList = new ArrayList<>();
+            webTemplateNodeList.add(webTemplate.getTree());
+            for (NodeId nodeId : nodeIds) {
+                webTemplateNodeList = webTemplateNodeList.stream()
+                        .map(n -> n.findMatching(f -> {
+                            if (f.getNodeId() == null) {
+                                return false;
+                            }
+                            // compere only classname
+                            else if (nodeId.getNodeId() == null) {
+                                return nodeId.getClassName().equals(new NodeId(f.getNodeId()).getClassName());
+                            } else {
+                                return nodeId.equals(new NodeId(f.getNodeId()));
+                            }
+                        }))
+                        .flatMap(List::stream)
+                        .collect(Collectors.toList());
+            }
+
+            final String aql;
+            Set<String> uniquePaths = new TreeSet<>();
+            webTemplateNodeList.stream().map(n -> n.getAqlPath(false)).forEach(uniquePaths::add);
+            if (uniquePaths.size() == 1) {
+                aql = uniquePaths.iterator().next();
+                //FIXME see https://github.com/ehrbase/project_management/issues/377
+            } else if (webTemplateNodeList.size() > 1) {
+                aql = uniquePaths.iterator().next();
+                log.warn(String.format("Aql Path not unique for template %s and path %s ", templateId, nodeIds));
+            } else {
+                aql = null;
+            }
+            if (aql != null) {
+                jsonPathQueryResult = new JsonPathQueryResult(templateId, aql);
             } else {
                 //dummy result since null can not be path of a cache
                 jsonPathQueryResult = new JsonPathQueryResult(null, Collections.emptyMap());
@@ -412,7 +480,50 @@ public class KnowledgeCacheService implements I_KnowledgeCache, IntrospectServic
 
             return null;
         }
+    }
 
+    @Override
+    public ItemInfo getInfo(String templateId, String aql) {
+        TemplateIdAqlTuple key = new TemplateIdAqlTuple(templateId, aql);
+        ItemInfo itemInfo = fieldCache.get(key);
+        if (itemInfo == null) {
+
+            WebTemplate webTemplate = getQueryOptMetaData(templateId);
+            String type;
+            Optional<WebTemplateNode> node = webTemplate.findByAqlPath(aql);
+            if (node.isEmpty()) {
+                type = null;
+            } else if (node.get().getRmType().equals("ELEMENT")) {
+                //for element unwrap
+                type = node.get().getChildren().get(0).getRmType();
+            } else {
+                type = node.get().getRmType();
+            }
+            String category;
+
+            if (node.isEmpty()) {
+                category = null;
+            } else if (aql.endsWith("/value")) {
+                //for element unwrap
+                category = webTemplate.findByAqlPath(aql.replace("/value", "")).filter(n -> n.getRmType().equals("ELEMENT")).map(n -> "ELEMENT").orElse("DATA_STRUCTURE");
+            } else {
+                category = "DATA_STRUCTURE";
+            }
+
+            itemInfo = new ItemInfo(type, category);
+            fieldCache.put(key, itemInfo);
+        }
+        return itemInfo;
+    }
+
+    @Override
+    public List<String> multiValued(String templateId) {
+        List<String> list = multivaluedCache.get(templateId);
+        if (list == null) {
+            list = getQueryOptMetaData(templateId).multiValued().stream().map(webTemplateNode -> webTemplateNode.getAqlPath(false)).collect(Collectors.toList());
+            multivaluedCache.put(templateId, list);
+        }
+        return list;
     }
 
     @Override
@@ -420,37 +531,5 @@ public class KnowledgeCacheService implements I_KnowledgeCache, IntrospectServic
         return this;
     }
 
-    /**
-     * Check an input byte array for a template if it is a valid template and generate a new template instance for it.
-     *
-     * @param templateContent - Byte array with template content
-     * @return - New instance of an OPT
-     */
-    private OPERATIONALTEMPLATE parseTemplate(byte[] templateContent) {
 
-        InputStream inputStream = new ByteArrayInputStream(templateContent);
-
-        TemplateDocument document;
-        try {
-            document = TemplateDocument.Factory.parse(inputStream);
-        } catch (XmlException | IOException e) {
-            throw new InvalidApiParameterException(e.getMessage());
-        }
-        OPERATIONALTEMPLATE template = document.getTemplate();
-
-        if (template == null) {
-            throw new InvalidApiParameterException("Could not parse input template");
-        }
-
-        if (template.getConcept() == null || template.getConcept().isEmpty())
-            throw new IllegalArgumentException("Supplied template has nil or empty concept");
-
-        if (template.getDefinition() == null || template.getDefinition().isNil())
-            throw new IllegalArgumentException("Supplied template has nil or empty definition");
-
-        if (template.getDescription() == null || !template.getDescription().validate())
-            throw new IllegalArgumentException("Supplied template has nil or empty description");
-
-        return template;
-    }
 }